requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.1
GDAL @ file:///C:/Users/Ricar/Downloads/GDAL-3.9.2-cp312-cp312-win_amd64.whl#sha256=83409fa04412efd11652601ddab443a1e255492c58eb5a98f001f5db3b60be3d
gunicorn==23.0.0
packaging==24.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
waitress==3.0.1
whitenoise==6.8.2
